Implement export and import json
