--- src/Files.App/Data/Settings/Settings.App.cs
@@ -1,16 +1,0 @@
-// Copyright (c) Files Community
-// Licensed under the MIT License.
-
-namespace Files.App.Data.Settings;
-
-public sealed partial class Settings
-{
-	[GeneratedSettingsProperty(DefaultValue = true)]
-	public partial bool ShowStatusCenterTeachingTip { get; set; }
-
-	[GeneratedSettingsProperty(DefaultValue = true)]
-	public partial bool ShowBackgroundRunningNotification { get; set; }
-
-	[GeneratedSettingsProperty(DefaultValue = false)]
-	public partial bool RestoreTabsOnStartup { get; set; }
-}
